cloud_storage: narrow the return type of get_materialized_manifest
The wide variant is used only in the cursor.

Motivation is that I want to use this method in other code
directly/without a cursor so I don't want to deal with unrelated variant
types which do not make sense there (monostate and stale manifest—can
ever exist only in the cursor).