ct: don't construct reconciler in data plane
The reconciler is not used anywhere currently, and it makes it tough to
wrangle dependencies. It depends on the cloud_topics::app that depends
on the data plane, which owns the reonciler, which is a logical circular
dependency.

Subsequent commits will iron out the dependency. For now this just
removes the reconciler from the data plane.