ct: reorganize cloud_topics::app
This reorganizes cloud_topics::app so that:
1. it is a single non-sharded entity,
2. it includes all cloud_topics-related sharded services

Previously, when cloud_topics::app was sharded, #2 was difficult because
there are some cloud_topics-related services that need to be sharded,
but app itself was sharded in application.cc; having a sharded service
own another sharded service seems like a bad idea.

In doing this, I needed to wrap the data_plane_api in another class in
order to shard it, which I've done with the recently added
state_accessors. This lets us continue to use the data_plane_api instead
of the data_plane_impl, as we were before, as a sharded service.

Using the state_accessors also lets us plumb the data plane API through
to cluster::partition and to the partition_proxy without introducing a
circular dependency. Previously we plumbed down cloud_topics::app, which
made it easy to trip over dependencies.