ct: introduce state_accessors
This introduces a wrapper around the data plane API, roughly the current
equivalent to cloud_topics::app. This will contain just state accessors
(data plane, eventually the metastore), rather than other long-lived
processes (reconciler, frontend, etc).

This is in preparation for cloud_topics::app to be just a container for
sharded services, which state_accessors will be one of.