cloud_io: Wire up client_lease_timeout_ms
Replace timeout computed from the incoming rtc node (high variance leading to
frequent timeouts) w/ the cluster config.

The overall goal is to avoid lease timeouts in the common case. We are generally
fine with uploads taking a long time, and the knock-on effects of frequent
timeout errors are not well understood, not to mention creating a lot of noise
in production cloud alerts.

Signed-off-by: Oren Leiman <oren.leiman@redpanda.com>

cloud_io: Use lease timeout cluster config

Replacing static factor of segment_upload_timeout_ms

Signed-off-by: Oren Leiman <oren.leiman@redpanda.com>