dt/cloud_retention: Use progress-oriented wait on KgoProducer
We've seen this slow way down in CDT but never stop making progress. Rather
than cranking the timeout even further, enable progress checking when we
wait_for_acks.

Note that we do increase the overall timeout from 10m to 20m, but this is
purposefully over-generous. If we actually stop making progress for some
reason, the underlying waiter should detect that.

Signed-off-by: Oren Leiman <oren.leiman@redpanda.com>