pbgen: pass in context to rpc methods
In order to support proxying API calls to the correct leader, we want to
ensure that we don't have proxy loops or other kinds of bad behavior. In
order to prevent this, we'll add a context call to every RPC method that
will tell handers two things: if this node has already seen this API,
and it will also hold other information like if this is a proxied
request, which will be needed for scatter-gather RPCs that need to
collect information from all over the cluster.

Secondly, while the context is being passed in we also update the RPC
handler to have a simpler interface. This allows it to be reused in
non-HTTP scenarios (like when we proxy using our internal RPC
interface), and we can push more of the HTTP handling out of codegen and
instead to hand written code in the admin server.