dl/coordinator: don't purge data when dropping Glue table
Glue doesn's support the purgeRequested option when dropping tables, and
will reject our requests indefinitely when a tombstone exists. This ends
up blocking topic tombstone removal, and therefore blocks subsequent
Iceberg topics of the same name from translating any data.

This commit plumbs a bool to the drop_table() command that is false for
Glue.

I considered doing this in the catalog layer, but opted to be more
explicit (it's hopeflly less surprising for readers that way).

Manually tested this against Glue:
1. create an Iceberg topic
2. produce to the Iceberg topic and wait for Iceberg commits
3. drop the Iceberg topic
4. witness the drop table op eventually succeeding in Glue
5. manually delete the data in Iceberg
6. recreate the topic and produce without issues

...as well as another sequence where I didn't do step 4 and instead
immediately went on to delete data and recreate the topic. In both
cases, the Redpanda coordinator was not blocked and could produce to the
new topic without issues.