.clangd: ignore -ferror-limit
You might reasonably put -ferror-limit=1 in your user.bazelrc, but that would
cause clangd to also use that flag, which is annoying because it means
that broken code won't be highlighted after the first error: but often
the first error may be way in some system header file, not even in the
file you are editing (in particular because clangd seems to have a
more expansive definition of error than clang).

So remove this from clangd's compile flags.