rptest: always decode backtraces when a test fails
When a test fails, we want to decode the any "seastar style" backtraces,
which then end up in redpanda_backtrace.log in the uploaded artifacts.

However, tests can fail in multiple ways, and in particular not all
failures occur in the test method: many occur in the @cluster wrapper
post-test checks. In some of these cases we did not decode the
backtrace, so we would not have the decoded backtrace in the artifacts.

This patch refactors the @cluster post-test checks into a separate
method, so that the code can be shared between the failure a non failure
paths, and moves the decoding such that it happens if either the test
fails or any post-test check fails.

Fixes CORE-12963.