ct/reconciler: Build objects by metastore object assignment
This is a large refactor of the reconciler to use the metastore to
drive object construction and register objects in L1.
1. The reconciler processes attached leader partitions in groups based
   on common object assignment by the L1 metastore. If the object grows
   too large, some partitions are left to be reconciled in a later run.
2. The reconciler commits the resulting objects to the L1 metastore and
   updates its LRO for the partition based on its work, or based on the
   metastore's corrected offset if the metastore provides ones.
This completes the integration of L1 with the reconciler.

Unfortunately, it's basically a rewrite of the core reconciliation
logic, and it's difficult to split apart because the metastore is
involved everywhere, and good overall organization developed organically
as I rewrote the reconciliation logic.

The structure of the code makes it amenable to unit testing but I've
left that out for now. The (ignored) ducktape test still passes, which
indicates that the reconciler still uploads something to the right
prefix. This is a basic indication that object construction and the
metastore operations succeed, but further testing is needed in
follow-ups.