ct/l1: add length of partition data to seek results
Previously, it was assumed we'd read the whole object, and just needed
the seek offset, now we return the length of the rest of the partition
data.

Thinking a bit more about this, I'm not actually sure this is what we
want due to caching and L1 objects. Right now the IO abstraction
transparently caches behind the scenes, but if we provide many different
lengths we'll cache a bunch of different ranges of the same object. We
probably need to change the IO abstraction so we can do a better job
caching.