k/client: Add support for DescribeCluster API
The DescribeCluster API carries the same information as the Metadata
request, however starting with Metadata v11, the cluster level
authorized operations were deprecated and moved to the new API. We still
want to use the most recent Metadata to get the topic IDs.  So this
update adds support for using the DescribeCluster API, if present, or
use Metadata v10 to get the cluster authorized ops.  Then a second
Metadata request is sent with the topics using the most up-to-date API.

As Redpanda currently supports DescribeCluster, we are able to test both
brokers with and without this API by simply not updating the
cluster_mock to support the DescribeCluster request.

Signed-off-by: Michael Boquard <michael@redpanda.com>