model: introduce new batch builder API
For a long time we've lived with using `storage::batch_builder` all over
the place, however really if all the logic for reading/appending to
batches is in model/ then we should have a builder there too. Also the
storage builder API is really only for internal consumers and make some
assumptions about the data used. Instead we define a new batch builder
interface that can be very simple for the normal use cases, but also
supports the full range of setting timestamps, etc for creating batches.

This is intended to be used in cloud topics to create placeholder
batches correctly.