ct/l1: make metastore::get_or_create_object_for return expected
It's possible that calls into the metastore will race with the topic
being created. One such case is in get_or_create_object_for().
Previously this always assumed the metastore topic was created, an
dereferenced an optional, resulting in undefined behavior.

This commit fixes this by exposing the error with an expected.