datalake: clean up files after translation error
We previously didn't clean up files record_multiplexer::finish()
returned errors. Since the returned success result is previously what
held the local files, this commit updates the signature to collect the
files in finish() (regardless of whether the operation succeeds or not),
and cleans up the files in translation_task::finish() and ::discard().