dt/cs_stress: Move status endpoint check to the epilogue
Similar motivation to 337e54dc87fab1b33b82ee0dcc43aa6fcf0e499b

Basically the test is one big race as written. Injecting some delay
between bucket view fetch and hitting status demonstrates this reliably.

The next best thing is to check that the manifest matches the bucket
at the end of the test.

Note that, after this change, most of the infrastructure in
cloud_storage_timing_stress_test will be intert. The async check runners
and all that. That's not ideal, but this commit represents the minimum
reasonable change to get the test not flaking (short of removing it
completely)

Signed-off-by: Oren Leiman <oren.leiman@redpanda.com>