tuner: Add infra for split live/dead IRQ assigment
This patch completes support for seaparate IRQ assignment of live and
dead IRQs (because of RX queue lowering).

While this was already partly implemented from perf_tune it was still
lacking some key functionality.

For now this code was (almost) never used and also didn't matter that
much in MQ mode.

However in dedicated mode with drivers that expose separate RX/TX IRQs
(like virtio) or those that don't remove names from /proc/interrupts
after RX-queue lowering (like virtio) this is needed to make dedicated
mode work correctly.

Hence port over the rest of the logic, the key bit is to use the queue
index to calculate the cut off point.

For a complete explanation of what this is all about please see the
comment I added.

In a follow up patch series we will add support for virtio and gvnic so
this is the predecessor for that.