security: use sasl mechanism_name consistently
Align the `{mechanism}::name` and `sasl_mechanism::mechanism_name()` for
consistency. This was consistent for most mechanisms except for SCRAM
and PLAIN.

`{mechanism}::name` is the kafka-official name, so align on that.

`sasl_mechanism::mechanism_name()` has only had two use cases: (1) used
as a string for pattern matching for audit logging, (2) used for
revoking OIDC credentials. (1) is fixed up in a backward-compatible way,
and (2) is unaffected.

This helps reuse the `{mechanism}::name` strings for pattern matching
into an enum for the `ListKafkaConnections` endpoint.