tuner: Skip some CDT tests on arm
x86 and arm assignment is different because of how SMT siblings are
layed out.

Just skip them on arm if so. We are not gaining much by testing on both
arm and x86.

I don't want to make it fully generic (i.e.: just calculate it with
hwloc) to have expectations fairly clear and avoid cyclical bugs.