cl/replication: Adjust start/stopping of replicators
With enabling mirroring of the _schemas topic, it is now required to
allow for a way to start replicators on a partition outside of
leadership change.  The _schemas topic may already exist when it becomes
a shadow topic and leadership for it may be stable.  This change permits
the ability to start replication on the _schemas topic.

There are now two lists that will contain requests to start or stop
replicators.  As long as the terms are increasing, stops remove starts
and vice versa.  A reconciler waiting on a condition variable will be
notified of changes and will enqueue the start/stop actions.

Signed-off-by: Michael Boquard <michael@redpanda.com>