dc/fetcher: invalidate epoch on offset reset
Currently offset_reset can result in an infinite loop of list_offset
requests without the offset actually getting reset when used along with
fetch_sessions.

The issue is that in the same iteration of do_fetch(), the fetcher may
choose to reset the offset but the fetch response may contain
offset_out_of_range error from the session un-doing the reset operation.

The fix is to reset the epoch along with the offset so the
offset_out_of_range from the same iteration is discarded.