dt: multiple CG connections list_kafka_connections
There are actually two connections in the response with the consumer
group id set. One is used by rpk to find the consumer group coordinator,
and the other one is the franz-go connection used for Join/SyncGroup.

This makes the test more robust by searching for the connection in the
list that matches the full predicate we expect.