cluster/epoch_service: fix inverted boolean
This was a straight bug that the bool was inverted. This caused us to
return invalid errors, or in success cases return a
std::unexpected(success), which are both wrong. The result of this is
that stale or invalid (-1) epochs could be returned from shard0.
Add an assertion to prevent the invalid epoch from being returned, and a
test that trips the assertion pre-fix.