cl/replication_mgr: fix replicator shutdown delays
With 9ce4feac38, the replicator start() is not trivial and involves
retries connecting to the source cluster to fetch the desired start
offset. The way shutdown is structured in the mgr, the gate is held
until the replicator has fully started or given up, this was causing the
shutdown delay as the replicator start is not preempted at shutdown.

This commit ensures the replicators are preempted during shutdown.