ct/reconciler: lie about skipped control batches
In the case of transactional control batches, we will skip them when
adding them to the object. However we need to give the metastore the
exact next offset to satisfy it. This requires that we lie to the
metastore and override the first observed offsets with the start offset
we attempted to reconcile from instead - as we're explicitly expecting
to skip offsets in this case.