tuner: Fix bond tuning to actually tune all slaves
There was a bug in the net_tuner which when tuning a bond interface
would only tune the first of possibly many slaves (the checkers were
already doing this correctly).

In practice this might not have mattered depending on how interfaces
were passed to the tuner and depending of the outcome of NIC auto
detection. Though in the most common cases the bug would have taken
effect.