tuner: Avoid repeated slave parsing
Further from the previous patch unroll the bond slaves initially which
allows us to effectively not deal with bond interfaces later as we are
only dealing with hardware interfaces then.

This was already implemented to a degree using `isSet` and
`tuneInterface` but like this we can get rid of it in even more places
and this avoids forgetting the filtering in the sub tuners.