cl/tests: deflake test_starting_offset
The issue is that if we do not wait for data to be replicated, the
consumer that probes for offset sees a missing offset (-1) in list_offset
responds and tries fetching from ts=-1 (latest_offset) and that resolves
to an incorrect offset (lso at the time of request).

The fix is to wait for hwm to advance (data is replicated) before it
queries for a timestamp.