s/appender: fixed race condition when waiting for flush operation
When executing `flush()` segment appender verifies if there are pending
bytes in head and instead of calling flush on the underlying `file` it
dispatches backgroud head write first and waits for flush completion
after that is a result of that background write and advancement of
committed offset.

Previously the appender called the `dispatch_background_head_write()`
method before accessing the future. In some rare situations it might
happened (especially when all the futures that are part of the
background flush are ready) that the operation was already completed and
deleted before its future was returned to the caller.

Now the future is requested first before returning to the caller this
way there is a guarantee that we are not accessing dangling reference.

Signed-off-by: Michał Maślanka <michal@redpanda.com>