treewide: remove bare excepts (E722)
Change all bare `except:` statements to catch `Exception` instead,
except where the exception is being unconditionally re-raised.

This makes `ruff check --select E722` pass cleanly, but specifically for
DT this is important as Type II test timeouts will get a
TestTimeoutError injected into the test, which inherits from
BaseException and would be caught by a bare except, but not by except
Exception. We don't want to catch this and eat them as it will prevent
this forced termination from working correctly. This is the same
general reasoning as to why one should not catch BaseException (or
use `except:`) in the first place.

Cases which unconditionally re-raise the exception are left as bare
excepts, since arguably here you want the logging or cleanup they do.
These are also not flagged by ruff.