cl/replicator: demote lack of monotonicity assert into an exception
This was originally an assert so we could catch things quickly during
development. A single partition running into this doesn't seem like a
reason to crash the broker. This commit changes it to a runtime
exception so the replicator can recover on a new leader and continue.

Do note that we still log an ERROR, so we are not losing any test
coverage. What was originally an assert will not be caught as a
BadLogLine.