lsm/block: introduce block builder
This commit introduces the block format and a builder class for it.
This is a very similar format to LevelDB, with some minor changes:

We use fixed width integers instead of varints when encoding. This
simplifies the implementation a bit and I don't think we care about the
space amplification. If we do decide we want the varints, then we can
introduce them later on and profile the differences.