lsm/db: introduce memtable
A memtable is the buffer in which data is accumulated before being
flushed to a L0 file. This implements a very similar read API as
SST files and really is just an initial "level" of files that can be
written too and we will sort in memory. We use an absl::btree to keep
the batches sorted. LevelDB uses skiplists, but since we're single core
only, btrees I think are a better data structure to use and it helps we
have a highly optimized implementation in absl we can use.

LevelDB Links:
* https://github.com/google/leveldb/blob/main/db/memtable.cc