lsm: add public interface to the database
The LSM database in this package is without a WAL (we rely on raft logs
for that), and we also hide the implementation and translate from
kafka/public API into internal API of the database.

The main motivation for splitting this core from a shell with the public
API surface, is to eventually supported a "tiered" database where we
stitch two LSM trees together. Why would that be useful? To support one
LSM tree on local disk and another in object storage, just like our
tiered storage implementation. With this, we can use a merging iterator
between the two databases to give a single database abstraction, as long
as we ensure that the object storage tier only has older data than the
local tier and there are no gaps (overlap would be fine however).
Anyways, that will not be implemented anytime soon, but the design
allows for it.

It's also an added benefit to be able to clearly see how kafka stuff
maps onto the LSM tree (offset -> seqno, etc).