kafka: catch shutdown errors in fetch
Following [1], make_reader() may throw if the underlying log is closed
while processing the request (rather than asserting). As is, the throw
in the handler causes the client request to time out. Instead, this
updates the shutdown case to return an error code indicating that we
are no longer leader (which seems reasonable if the partition is
shutting down).

1. https://github.com/redpanda-data/redpanda/pull/28328