k/tests: make write at offset stm test more strict
Changed write at offset stm test to validate the offsets on each replica
after the snapshot is applied. Previously the tests was not generating
any Raft batches so the offset translator state didn't influence the
test outcome.

Now the snapshot recovery test includes the offset translator handled
batches to force the offset translator state to be non trivial.
Additionally the test now validates the expected start offset on all of
the replicas

Signed-off-by: Michał Maślanka <michal@redpanda.com>