ts: Don't log shutdown error on 'error' level
Previously, it was assumed that only the actual errors are propagated to
the remote partition reader via the log reader config. This is actually
not the case and the kafka layer can propagate shutdown errors too. This
leads to the correct behavior. The current async operation stops due to
shutdown exception but the exception is not propagated further. The
kafka layer propagates 'ssx::shutdown_requested_exception' which is
derived from the 'ss::abort_requested_exception' so the catch block
inside the 'partition_record_batch_reader_impl::do_load_slice' can
handle it. The 'throw_on_external_abort' also converts any connection
related error to 'ss::abort_requested_exception' so it is definitely
safe to allow this exception to be propagated.

The commit changes the log level and the message and removes the comment
which is no longer correct.

Signed-off-by: Evgeny Lazin <4lazin@gmail.com>