ts: Fix cache trim race
The race happens when the fast trim can't remove enough data and starts
full trim. If while the fast trim was running the access time tracker
was snapshotted the updates generated by the fast trim may not be
immediately available for reading because they will be added to the
_pending_upserts collection.

If the next trim is started when this collection is not empty the trim
may try to remove files which are already removed. The method
'lru_entries' previously didn't take '_pending_upserts' into account.

This commit futurizes 'lru_entries'. The method first acquires units and
then returns the result. When the units are acquired it's guaranteed
that the collection is empty and 'lru_entries' will return up to date
results.

Signed-off-by: Evgeny Lazin <4lazin@gmail.com>