rptest: fix BadLogLines exception in __str__
Maybe. In OMB we pass the wrong type to BLL which causes its __str__
to throw, which confuses DT which doesn't get any stack trace and
just prints "Test Passed" (but the test fails).

Also change the BLL __str__ method so it cannot throw. We create
the string up front in the constructor, which will throw if it fails,
still in the test context so will be reported properly by DT.

Improve types a bit in this area (Stephan has another change which does
this more broadly for omb benchmark), since typing would have caught
this when the bug was originally introduced.

Fixes CORE-13453.