rptest: metrics simplification, typing
Simplify metrics implementation in RedpandaService[Cloud] and associated
classes by removing redundant wrappers in RS and RSC and doing almost
all work in the ABC. The derived classes just need to provide a list
of nodes `get_nodes_abc()` as they vary there (nodes vs pods) and also
the actual metrics query method.

This makes it less painful to make changes to metrics and reduces the
chance of errors on the cloud side which is harder to test.

Also, improve the typing in this area.