rptest: introduce targeted metrics query
Seastar allows filtering on the metric name by passing `?__name__=foo`
to the endpoint. Use this in ducktape `metrics_sample` and
`metrics_samples` in order improve metrics query performance, as
serializing a huge response (approaching 100MB in tests with high
numbers of partitions) only to throw away 99.999% of it is slow.

We implement this using a name `name`/`names` parameter to
metrics_sample/samples. If set, we use the "exact" query. We can't
easily convert all existing calls to this mechanism since the existing
semantics do a fuzzy search which is not supported by the seastar
metrics endpoint.

We require one query per metric name, so if you pass _many_ names, it
may be more efficient to use the existing mechanism. Most uses pass one
name, however.

Relates to CORE-13166.