c/data-migrate/router: use controller abort source
`cluster::data_migration_router` is a wrapper on top of a couple of
instances of `cluster::leader_router`.

`cluster::leader_router` will search for partition leaders and direct
requests to appropriate nodes. If leader is not found it'll backoff and
retry, with current settings for up to 20 minutes altogether.

This may lead to slow shutdowns. Abort early.