dt: Relax partition HWM check
Due to a bug in fetch sessions, sometimes the source HWM is not returned
in a fetch response.  This occurs primarily if the topic is not active
and there is no more data to read after a partition is added to the
fetch session.  See CORE-14617 for more information.

Filed CORE-14653 to return to this and re-enable the checks.

Signed-off-by: Michael Boquard <michael@redpanda.com>