cloud_storage: fix request signing with special chars
We found an issue where continuation token contains chars which look
like url encoded, e.g. plus sign (+). If we send it as-is in the query
param but then we url-encode it in the signature the server responds
with invalid signature error. That's because it (aws server)
unconditionally decodes the query parameters first.

Even if we fixed the request signature it is like that listing won't
work as the `+` sign must be sent as `+` and not let the server decode
it as a space instead!

So we need to encode query params. This PR does that. And now, we don't
need to url encode during signing!