c/migrations/tx_manager_migrator: use ungated c/leader_router
RPC handlers are not designed to have an async stop(), but one of them,
`tx_manager_migrator_handler`, uses leader routers. It is safe, as RPC
server is drained before the handlers are destroyed.

Introduce an ungated flavour of cluster::leader_router for such use.
It's also going to be used in tx_manager_migrator itself, as it uses
the same leader_router subclasses. It's also safe, as its use is led by
Admin API, which is drained before tx_manager_migrator.