`cloud_storage_clients`: fix partial `xml_sax_parser` reads
We occasionally see CI failures due to what seem to be partial
reads of XML elements. This is because we have erroneously assumed
characters in our `handle_characters` callback to `xmlSAXHandler`
always contained a full read of an XML element.

Fix partial reads in `xml_sax_parser` by continually appending
to a `_current_chars` string that is eventually consumed within
`handle_end_element()` in parser implementations.