rptest/typing: add a few additional checks
Add some additional checks which aren't turned on in strict mode:

reportPropertyTypeMismatch
reportUnnecessaryTypeIgnoreComment

Removed a few # type: ignore comments which are no longer needed

reportUnreachable

Probably the most quality-improving check in this batch. Fixed all
existing cases of this warning. Many cases would ideally use
`assert_never`, but this doesn't work as well you'd like due to
https://github.com/microsoft/pyright/issues/11120, but that is used
in some places, and in others we just surpress the error (typically this
happens when some if/else or match cases are exhaustive and we have
an "unexpected value" case as the default).

In other cases code was truly dead and was deleted and in others the
code was live but the typing above it was wrong, in these cases I
fixed the typing. One such case was ping_pong.py and there I typed the
entire file.