lsm/io: introduce cloud persistence
Cloud persistence is basically the same model as RocksDB Cloud, except
we use CAS primatives to ensure only a leader is writing to the
database. All writes are staged locally, then uploaded to the cloud.
Likewise, reads are first downloaded to the local disk, then served
locally. This scheme allows for a recovery option if the cluster is
destroyed, the data can still be recovered from object storage.