rptest: switch to ripgrep in log search
Use ripgrep (rg) instead of grep for log searching in rptest.

In my benchmark this results in about a 3x speedup when searching
through 3 logs in parallel resulting from creating 10 topics of 1000
partitions.

The benchmark is also included in this change, though @ignored so it
does not run CI (it takes a minute or so).

This requires translating GNU BRE (used by grep by default) to ERE
(used by grep -E, and ripgrep). It was probably a mistake to use BRE
in the first place, but it is what it is.