http: consume body for HEAD responses
Although the HTTP RFCs require servers not to send any body in the
responses we can't assume that all servers are compliant. So we are
defensive and always _parse_ the body. However, we instruct boost::beast
to skip over it. This guarantees a reusable client/connection.

The actual driver is not non-compliant servers but rather leaving
boost::beast parser in a valid state, always. Before this change, after
handling HEAD responses, the parses would have been left in an
incomplete state.

This change unlocks an improvement of the resilience of our http client
in face of bugs which we'll implement next. I.e. invalidate clients left
in an _improper_ state.