configuration: added default rpc timeout property
Added a configuration property intended to unify the RPC timeout
management throughout the application. The property is intended to
replace some existing timeout configuration variable and be a default
for future RPC requests and services. Other timeouts f.e. max value of
backoff timeout may and should be derived based on this property.

Signed-off-by: Michał Maślanka <michal@redpanda.com>