cloud_storage_clients/abs: Drop showonly=files when using a delimiter
When listing blobs on an Azure account with HNS enabled, the client was
adding showonly=files to the list objects request even when using a
delimiter. However, the inventory report was using this request to
discover virtual directories, so showonly=files broke it. This caused
the CloudStorageInventoryTest.test_load_inventory_report to fail on
Azure with HNS enabled.

The fix is to remove showonly=files when a delimiter is specified,
because the purpose of the delimiter is to group nested objects into
prefixes, and showonly=files negates that.

Note that the Azurite container used in ducktape to mock ABS doesn't
behave exactly the same as ABS with HNS enabled, so this test works fine
when run against mock ABS using Azurite.

Fixes CORE-8225