kgo-verifier: update franz-go to always retry EOF
Tests are flaking flake when kgo-repeater dies with a
non-retriable error right after connection. This happens when franz-go
gets EOF right after connecting but before receiving any response, as
a heuristic it assumes this is a SASL misconfiguration as that is the
broker behavior in that case. However, this can also occur because
Redpanda is stopped/killed after the connection is made but before
the initial requests can be responded to.

This means the producer will fail if a producer dies/is killed/stops
during a critical window between the connection and receiving the first
response. This is reasonably likely in stress/chaos tests where
producers are being started and stopped all the time.

This is a relatively recent change (~6 months ago) in franz-go,
which was brought in a few months ago by a franz-go upgrade.

To fix this, we make use a proposed new option to franz-go, from this
PR:

https://github.com/twmb/franz-go/pull/1198

This is not merged, so we pull in the SHA from this PR directly. When
a franz-go release is made with this change, we can update to that.

Details in CORE-14849.

Fixes CORE-14898.