cluster/partition: Add timequery empty range check to handle empty log
It can happen that a timequery hits an empty local log and its max
offset is before the min offset, which is clamped to the start of the
log. This will fail translation. This change adds extra early returns to
local and cloud timequeries to handle this special case pre-translation.