`ct/l1`: soften requirements in `replace_objects()` below start offset
It is not hard to imagine a concurrent race between a `replace_objects()`
and a `set_start_offset()` update to the `metastore` (e.g. a race between
compaction of a log and a prefix truncation).
Both outcomes of the race are at the present time handled correctly:

1. If `replace_objects()` lands first, `set_start_offset()` will simply
   discard whatever extents are in the `metastore` below the new start offset.
2. If `set_start_offset()` lands first, the `replace_objects()` request will
   _likely_ be rejected (it may still be accepted if the offset update
   does not remove any existing `extent`s).

Case 2 is currently handled pessimistically. We do not need to reject
a `replace_objects()` command just because some of the objects in the update
are below the new start offset- it does not affect the validity of the objects
above the new start offset.

Handle this case better by tweaking the logic in `get_range()` to ignore
extent alignment requirements for replacements below the log start offset,
and then remove the un-needed extents after they have been added in `apply()`.

We allow for misalignment between the start offset and the first extent in
the partition, since we need to support partial truncations- this also means
we don't need replacement extents to be aligned to existing extents below the
start offset. We will still need to validate the alignment of the range's
last offset to be sure the update is valid.