ct: add l0 object size distribution test
This test wires up both the write request scheduler and the batcher and
verifies that the batcher is making l0 object upload requests with
object sizes that match our expectations given a particular write
workload.

The intention is for this test to serve as a sort of regression test on
the effectiveness of L0 to produce "good" object sizes which are an
important component of L0 cost.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>